+++ NEWS.md
@@ -5,3 +5,5 @@
 # flowchart 0.2.0
 
 * Updated description with minimum R version (>= 4.1.0)
+
+* Older minimum version for dplyr